backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.4.0
billiard==4.2.4
celery==5.5.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.4
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.1
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.2
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-telegram-bot==20.6
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
twilio==9.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.22.0
